+++ src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/function/postgis/raster/attribute/NumXTiles.java
@@ -12,8 +12,8 @@ public class NumXTiles extends RasterAttributeFunction{
 	}
 
 	@Override
-	public double attribute(GridCoverage raster) {
-		return raster.render(raster.getGridGeometry().getExtent()).getNumXTiles();
+	public Double attribute(GridCoverage raster) {
+		return (double) raster.render(raster.getGridGeometry().getExtent()).getNumXTiles();
 	}
 
 }
